Fix -pedantic warnings
struct with char* initialization needs {0}.
-Wsign-compare, etc.

  * src/search.cc (~Search): fix -Wsign-compare.
  * tests/lang-ucs2.gperf: add %null-strings.
  * tests/lang-utf8.gperf: add %null-strings.
  * tests/smtp.gperf: change k to size_t. -Wsign-compare.
  * tests/test2.c (main): silence -Wunused argv.

+++ src/search.cc
@@ -1750,7 +1750,7 @@ Search::~Search ()
                  _min_key_len, _max_key_len);
 
       free ((void*)nbperf->keylens);
-      for (int i=0; i < nbperf->n; i++)
+      for (size_t i=0; i < nbperf->n; i++)
         free ((void*)nbperf->keys[i]);
       free ((void*)nbperf->keys);
